requirements.txt
fastapi
uvicorn[standard]
firebase-admin
google-generativeai
pydantic
python-multipart